+++ activestorage/app/controllers/active_storage/base_controller.rb
@@ -1,10 +1,11 @@
 # frozen_string_literal: true
 
 # The base class for all Active Storage controllers.
-class ActiveStorage::BaseController < ActionController::Base
+class ActiveStorage::BaseController < ActionController::API
+  include ActionController::RequestForgeryProtection
   include ActiveStorage::SetCurrent
 
-  protect_from_forgery with: :exception
+  protect_from_forgery with: :exception, if: :protect_against_forgery?
 
   self.etag_with_template_digest = false
 
@@ -16,4 +17,8 @@ class ActiveStorage::BaseController < ActionController::Base
     ensure
       response.stream.close
     end
+
+  public
+
+  ActiveSupport.run_load_hooks :active_storage_base_controller, self
 end