+++ activestorage/app/controllers/active_storage/disk_controller.rb
@@ -7,7 +7,7 @@
 class ActiveStorage::DiskController < ActiveStorage::BaseController
   include ActiveStorage::FileServer
 
-  skip_forgery_protection
+  skip_forgery_protection if: :protect_against_forgery?
 
   def show
     if key = decode_verified_key